phase-2-cloud-ai/financial-behavior-insights/requirements.txt
azure-ai-ml
azureml-mlflow
azure-core
azure-identity
azure-keyvault-secrets
azure-storage-blob
dvc
mlflow
scikit-learn
pandas
numpy
joblib
python-dotenv
pytest
flask
pydantic
seaborn
matplotlib
scipy
requests
dvc[azure]


